requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_JWT_Extended==4.4.4
Flask_Login==0.6.3
Flask_Mail==0.9.1
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
PyJWT==2.8.0
pytz==2024.1
SQLAlchemy==2.0.36
Werkzeug==3.1.3
